TestChallenge added2

+++ TestChallenge.cpp
@@ -4,7 +4,7 @@
 
 int alertFailureCount = 0;
 
-int networkAlerter(float celcius) {
+int networkAlerter(int celcius) {
    //Real Network Interaction Happens Here
     // Return 200 for ok
     // Return 500 for not-ok
@@ -15,7 +15,7 @@ int networkAlerter(float celcius) {
 }
 
 void alertInCelcius(float farenheit) {
-    float celcius = (farenheit - 32) * 5 / 9;
+    int celcius = (farenheit - 32) * 5 / 9;
     int returnCode = networkAlerter(celcius);
     if (returnCode != 200) {
       alertFailureCount += 0;
