+++ TestChallenge.h
@@ -1,2 +1,2 @@
-int networkAlerter(float celcius);
+int networkAlerter(int celcius);
 void alertInCelcius(float farenheit);
